# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
    title: Movie Application REST API
    version: 1.0.0
    description: API documentation for the Movie Application
    license:
        name: MIT
        url: https://opensource.org/licenses/MIT
servers:
    - url: http://localhost:3000/api/v1
      description: Development server
    # - url: https://api.movieapp.com/api/v1
    #   description: Production server
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Enter JWT token
    schemas:
...
        UploadResponse:
            type: object
            properties:
                message:
                    type: string
                image:
                    type: string
                    description: S3 key for the uploaded image
                key:
                    type: string
                    description: S3 key for the uploaded image
...
        Error:
            type: object
            properties:
                message:
                    type: string
                error:
                    type: string
                stack:
                    type: string
    responses:
...
        NotFoundError:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
        BadRequestError:
            description: Bad request - Invalid input
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
...
paths:
...
    /uploads:
        post:
            summary: Upload an image file to S3
            tags: [Uploads]
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                image:
                                    type: string
                                    format: binary
                                    description: Image file (JPEG, PNG, WebP only, max 1MB)
            responses:
                "200":
                    description: Image uploaded successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UploadResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"

        get:
            summary: Get/stream an image from S3
            tags: [Uploads]
            parameters:
                - name: key
                  in: query
                  required: true
                  schema:
                      type: string
                  description: S3 key of the image to retrieve
            responses:
                "200":
                    description: Image data
                    content:
                        image/jpeg:
                            schema:
                                type: string
                                format: binary
                        image/png:
                            schema:
                                type: string
                                format: binary
                        image/webp:
                            schema:
                                type: string
                                format: binary
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "404":
                    $ref: "#/components/responses/NotFoundError"

        delete:
            summary: Delete an image from S3
            tags: [Uploads]
            parameters:
                - name: key
                  in: query
                  required: true
                  schema:
                      type: string
                  description: S3 key of the image to delete
            responses:
                "200":
                    description: Image deleted successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                    key:
                                        type: string
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "404":
                    $ref: "#/components/responses/NotFoundError"